Rename homework3 package to searchengine
