+++ src/main/scala/searchengine/http/HttpResponse.scala
@@ -1,4 +1,4 @@
-package homework3.http
+package searchengine.http
 
 import java.nio.charset.{Charset, StandardCharsets}
