+++ src/main/scala/searchengine/http/HttpUtils.scala
@@ -1,4 +1,4 @@
-package homework3.http
+package searchengine.http
 
 import java.net.URI
