+++ src/main/scala/searchengine/processors/BrokenLinkDetector.scala
@@ -1,7 +1,7 @@
-package homework3.processors
+package searchengine.processors
 
-import homework3.Processor
-import homework3.http.HttpResponse
+import searchengine.Processor
+import searchengine.http.HttpResponse
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
